requirements.txt
absl-py==1.2.0
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.8.1
altair==4.2.0
anyio==3.6.1
apprise==1.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgi-lifespan==1.0.1
astroid==2.12.11
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
autopep8==1.7.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
Bottleneck==1.3.5
branca==0.5.0
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-binary==77.0.3865.40.0
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.5
coolname==1.1.0
croniter==1.3.7
cryptography==38.0.1
cycler==0.11.0
Cython==0.29.32
dash==2.7.1
dash-bootstrap-components==1.3.0
dash-colorscales==0.0.4
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2022.9.2
databricks-cli==0.17.3
db-dtypes==1.0.4
deap==1.3.3
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
docker==6.0.0
dtale==2.10.0
entrypoints==0.4
et-xmlfile==1.1.0
etils==0.8.0
exceptiongroup==1.0.1
executing==1.1.0
fastapi==0.85.0
fastjsonschema==2.16.2
filelock==3.9.0
flake8==5.0.4
Flask==2.2.2
Flask-Compress==1.13
flask-ngrok==0.0.25
flatbuffers==22.9.24
folium==0.12.1.post1
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.8.2
future==0.18.3
gast==0.4.0
gdown==4.6.0
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.28
google-api-core==2.8.2
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.34.4
google-cloud-bigquery-storage==2.16.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
google-trans-new==1.1.9
googleapis-common-protos==1.56.4
graphviz==0.20.1
greenlet==1.1.3
griffe==0.22.2
grpcio==1.49.1
grpcio-status==1.48.2
gunicorn==20.1.0
h11==0.12.0
h5py==3.7.0
htmlmin==0.1.12
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.12.0
idna==3.4
ImageHash==4.3.1
imageio==2.22.1
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==4.13.0
importlib-resources==5.9.0
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.16.0
ipympl==0.9.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.1
json5==0.9.10
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.11.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.5.0
jupyter-resource-usage==0.6.3
jupyter-server==1.21.0
jupyter_client==7.4.2
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.2
kaggle==1.5.12
kaleido==0.2.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kubernetes==24.2.0
lazy-object-proxy==1.7.1
libclang==14.0.6
locket==1.0.0
lxml==4.9.1
lz4==4.3.2
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
memoized-property==1.0.3
missingno==0.4.2
mistune==2.0.4
mlflow==1.29.0
mlxtend==0.21.0
multidict==6.0.2
multimethod==1.4
nbclassic==0.4.4
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.6.1
nbresult==0.0.8
nest-asyncio==1.5.6
networkx==2.8.7
nltk==3.7
nmap==0.0.1
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.1
openai==0.26.5
openpyxl==3.0.10
opt-einsum==3.3.0
orjson==3.8.0
outcome==1.2.0
packaging==21.3
pandas==1.4.4
pandas-gbq==0.17.9
pandas-profiling==3.3.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pathspec==0.10.1
patsy==0.5.2
pendulum==2.1.2
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.10.0
pluggy==1.0.0
pmdarima==2.0.1
prefect==2.5.0
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
promise==2.3
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==3.19.6
psutil==5.9.2
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.2
pydata-google-auth==1.4.0
pydeck==0.8.0b3
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.5.0
pylint==2.15.4
Pympler==1.0.1
pyparsing==3.0.9
pyreadstat==1.2.0
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.3
pytest-asyncio==0.19.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
querystring-parser==1.2.4
readchar==4.0.3
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
savReaderWriter==3.4.2
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.11.2
selenium==4.6.0
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.97
setuptools-scm==7.0.5
six==1.16.0
smart-open==6.2.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlparse==0.4.3
squarify==0.4.3
stack-data==0.5.1
starlette==0.20.4
statsmodels==0.13.2
stopit==1.1.2
streamlit==1.18.1
strsimpy==0.2.1
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-datasets==4.7.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-metadata==1.10.0
termcolor==2.0.1
terminado==0.16.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.8.12
tinycss2==1.1.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
toolz==0.12.0
torch==1.13.1
tornado==6.2
TPOT==0.11.7
tqdm==4.64.1
traitlets==5.4.0
transformers==4.26.1
trio==0.22.0
trio-websocket==0.9.2
typer==0.6.1
typing_extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
Unidecode==1.3.6
update-checker==0.18.0
urllib3==1.26.12
uvicorn==0.18.3
validators==0.20.0
visions==0.7.5
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
wsproto==1.2.0
xarray==2022.12.0
xgboost==1.6.2
xlrd==2.0.1
XlsxWriter==3.0.3
yapf==0.32.0
yarl==1.8.1
zipp==3.8.1
